Started on some lexer tests but I realize now there is probably a better way to resolve the regex overlaps.
