+++ futhark-tests/.gitignore
@@ -1,2 +1,3 @@
 *
 !.gitignore
+!futhark.pkg
